rename(container): 폴더명 수정(#219)
폴더명 수정에 따라서 import 경로를 수정하였습니다.
